requirementsV1Lockfilebackup.txt
flask                              # Web framework for routing and server logic
pandas                             # Data analysis and manipulation
requests                           # Makes HTTP requests (used for eBay API, etc.)
gspread                            # Interact with Google Sheets

google-auth                        # Core authentication for Google APIs
google-auth-httplib2               # HTTP support for google-auth
google-auth-oauthlib               # OAuth2 login support for Google APIs
google-api-python-client           # Full client for Google Sheets/Drive API

openpyxl                           # Read/write Excel .xlsx files
xlrd                               # Read legacy .xls files

gunicorn                           # Production WSGI server for Flask apps
python-dotenv                      # Load environment variables from .env file
jinja2                             # Templating engine used in Flask HTML
oauthlib                           # General OAuth protocol support



(venv) C:\Users\Justin\Desktop\SELL PROJECT\CODE\manifest_app_final_v1>python -m pip list
Package                  Version
------------------------ -----------
blinker                  1.9.0
cachetools               5.5.2
certifi                  2025.4.26
charset-normalizer       3.4.2
click                    8.2.1
colorama                 0.4.6
et_xmlfile               2.0.0
Flask                    3.1.1
google-api-core          2.25.0
google-api-python-client 2.172.0
google-auth              2.40.3
google-auth-httplib2     0.2.0
google-auth-oauthlib     1.2.2
googleapis-common-protos 1.70.0
gspread                  6.2.1
gunicorn                 23.0.0
httplib2                 0.22.0
idna                     3.10
itsdangerous             2.2.0
Jinja2                   3.1.6
MarkupSafe               3.0.2
numpy                    2.3.0
oauthlib                 3.2.2
openpyxl                 3.1.5
packaging                25.0
pandas                   2.3.0
pip                      25.1.1
proto-plus               1.26.1
protobuf                 6.31.1
pyasn1                   0.6.1
pyasn1_modules           0.4.2
pyparsing                3.2.3
python-dateutil          2.9.0.post0
python-dotenv            1.1.0
pytz                     2025.2
requests                 2.32.4
requests-oauthlib        2.0.0
rsa                      4.9.1
six                      1.17.0
tzdata                   2025.2
uritemplate              4.2.0
urllib3                  2.4.0
Werkzeug                 3.1.3
xlrd                     2.0.1